[AC-7712] one class per file
