+++ web/impact/impact/v1/serializers/location_serializer.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+
+from accelerator.models import Location
+
+
+class LocationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Location
+        fields = ['id', 'street_address', 'timezone', 'country',
+                  'state', 'name', 'city', ]
