+++ web/impact/impact/v1/serializers/user_serializer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from accelerator.models import User
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ['id', 'first_name', 'last_name', 'email']
